+++ examples/personalized_fl/algorithms/SSL/byol/byol_trainer.py
@@ -5,7 +5,7 @@ from lightly.utils.scheduler import cosine_schedule
 from lightly.models.utils import update_momentum
 
 
-from pflbases import ssl_trainer
+from ssl import ssl_trainer
 from plato.trainers import loss_criterion
 from plato.config import Config
