+++ examples/personalized_fl/algorithms/SSL/calibre/calibre.py
@@ -1,8 +1,8 @@
 """
 Implementation of our Calibre algorithm.
 """
-from pflbases import ssl_datasources
-from pflbases import ssl_client
+from ssl import ssl_datasources
+from ssl import ssl_client
 
 
 import calibre_model
