+++ examples/personalized_fl/algorithms/SSL/smog/smog.py
@@ -6,11 +6,10 @@ ECCV, 2022. https://arxiv.org/pdf/2006.07733.pdf.
 """
 
 
-from pflbases import fedavg_personalized
-from pflbases import fedavg_personalized
+from plato.servers import fedavg_personalized as personalized_server
 
-from pflbases import ssl_client
-from pflbases import ssl_datasources
+from ssl import ssl_client
+from ssl import ssl_datasources
 
 
 import smog_trainer
@@ -26,10 +25,9 @@ def main():
         datasource=ssl_datasources.SSLDataSource,
         trainer=smog_trainer.Trainer,
     )
-    server = fedavg_personalized.Server(
+    server = personalized_server.Server(
         model=smog_model.SMoG,
         trainer=smog_trainer.Trainer,
-        algorithm=fedavg_personalized.Algorithm,
     )
 
     server.run(client)
