+++ examples/personalized_fl/algorithms/configs/SSL/smog_CIFAR10_resnet18.yml
@@ -25,7 +25,6 @@ data:
     !include cifar10_ssl_noniid.yml
 
 trainer:
-
     # The maximum number of training rounds
     rounds: 2
 
@@ -59,7 +58,7 @@ trainer:
 
 algorithm:
     # Aggregation algorithm
-    type: fedavg
+    type: fedavg_personalized
     #  - STL10 (Image size=96)
     #  - CIFAR10 (Image size=32)
     #  - MNIST (Image size=28)
